// types/match.ts
export interface Match {
  team_b_id: number;
  date_wise: string;
  max_rate: string;
  match_id: number;
  venue: string;
  match_status: string;
  matchs: string;
  venue_id: number;
  series: string;
  team_a_id: number;
  match_date: string;
  team_a_img: string;
  min_rate: string;
  match_time: string;
  match_type: string;
  team_b_img: string;
  team_b_short: string;
  team_b: string;
  team_a_short: string;
  fav_team: string;
  team_a: string;
  is_hundred: number;
  series_id: number;
  series_type: string;
}